ccsms-backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0
drf-yasg==1.21.7
django-filter==23.5
faker==22.2.0
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0